Move validation component to different folder (/Forms)
